do clean up
