+++ library/src/getitune/backend/lightning/models/instance_segmentation/heads/__init__.py
@@ -4,6 +4,7 @@
 """Custom head implementations for instance segmentation task."""
 
 from .bbox_head import ConvFCBBoxHead
+from .ec_segmentation_head import SegmentationHead
 from .fcn_mask_head import FCNMaskHead
 from .roi_head import RoIHead
 from .roi_head_tv import TVRoIHeads
@@ -16,5 +17,6 @@ __all__ = [
     "RPNHead",
     "RTMDetInstSepBNHead",
     "RoIHead",
+    "SegmentationHead",
     "TVRoIHeads",
 ]
